New agent: Initial import of the Ansible based agent
Few months ago, a work to produce a new version of the DCI agent has
been started. The main goal being that this agent should be flexible (by
adding feature through plugins) and using Ansible as much as possible.

The following patchset is the creation of the new feature-branch
ansible. Goal is to unify the old-agent work with the new agent one.

Change-Id: Icc12a04586ad866d18ca8440783b9d9430a3e8de