Avoid namespace collision during import
The module name is 'dci_agent' the script name is 'dci_agent', this
result in namespace collision during import.

This make one uses workaround like from . import X because it can not
import correctly the wished module.

This commit fixes that by chaning the name of the script to shell.py
and demonstrating it works by importing version.__version__ from the
shell.py itself.

Change-Id: Icbdb9cf881d36035d3e32efd786a57fdb84d0c0c